requirements.txt
requests
gql
pytest
python-dotenv
